refactor(plugin-eslint): prepare different versions of loading rules
