+++ packages/plugin-eslint/src/lib/meta/groups.ts
@@ -2,7 +2,7 @@ import type { Rule } from 'eslint';
 import type { Group, GroupRef } from '@code-pushup/models';
 import { objectToKeys, slugify } from '@code-pushup/utils';
 import { ruleIdToSlug } from './hash';
-import { type RuleData, parseRuleId } from './rules';
+import { type RuleData, parseRuleId } from './parse';
 
 type RuleType = NonNullable<Rule.RuleMetaData['type']>;
