+++ packages/plugin-eslint/src/lib/meta/parse.ts
@@ -1,0 +1,33 @@
+import type { Linter, Rule } from 'eslint';
+
+export type RuleData = {
+  ruleId: string;
+  meta: Rule.RuleMetaData;
+  options: unknown[] | undefined;
+};
+
+export function parseRuleId(ruleId: string): { plugin?: string; name: string } {
+  const i = ruleId.lastIndexOf('/');
+  if (i < 0) {
+    return { name: ruleId };
+  }
+  return {
+    plugin: ruleId.slice(0, i),
+    name: ruleId.slice(i + 1),
+  };
+}
+
+export function isRuleOff(entry: Linter.RuleEntry<unknown[]>): boolean {
+  const level: Linter.RuleLevel = Array.isArray(entry) ? entry[0] : entry;
+
+  switch (level) {
+    case 0:
+    case 'off':
+      return true;
+    case 1:
+    case 2:
+    case 'warn':
+    case 'error':
+      return false;
+  }
+}
