+++ packages/plugin-eslint/src/lib/meta/rules.unit.test.ts
@@ -2,7 +2,8 @@ import { dirname, join } from 'node:path';
 import { fileURLToPath } from 'node:url';
 import type { MockInstance } from 'vitest';
 import type { ESLintTarget } from '../config';
-import { type RuleData, listRules, parseRuleId } from './rules';
+import type { RuleData } from './parse';
+import { listRules } from './rules';
 
 describe('listRules', () => {
   const fixturesDir = join(
@@ -158,36 +159,5 @@ describe('listRules', () => {
         options: [],
       } satisfies RuleData);
     });
-  });
-});
-
-describe('parseRuleId', () => {
-  it.each([
-    {
-      ruleId: 'prefer-const',
-      name: 'prefer-const',
-    },
-    {
-      ruleId: 'sonarjs/no-identical-functions',
-      plugin: 'sonarjs',
-      name: 'no-identical-functions',
-    },
-    {
-      ruleId: '@typescript-eslint/no-non-null-assertion',
-      plugin: '@typescript-eslint',
-      name: 'no-non-null-assertion',
-    },
-    {
-      ruleId: 'no-secrets/no-secrets',
-      plugin: 'no-secrets',
-      name: 'no-secrets',
-    },
-    {
-      ruleId: '@angular-eslint/template/no-negated-async',
-      plugin: '@angular-eslint/template',
-      name: 'no-negated-async',
-    },
-  ])('$ruleId => name: $name, plugin: $plugin', ({ ruleId, name, plugin }) => {
-    expect(parseRuleId(ruleId)).toEqual({ name, plugin });
   });
 });
