+++ packages/plugin-eslint/src/lib/meta/versions/detect.ts
@@ -1,0 +1,6 @@
+import type { ConfigFormat } from './formats';
+
+export async function detectConfigVersion(): Promise<ConfigFormat> {
+  // TODO: detect flat config
+  return 'legacy';
+}
