+++ packages/plugin-eslint/src/lib/meta/versions/formats.ts
@@ -1,0 +1,1 @@
+export type ConfigFormat = 'flat' | 'legacy';
